Fix challenge decode: seedValue Int->UInt64, add test verification button
The API returns seed_value as a uint64 that can exceed Int64 max
(e.g. 15232519584105517258 > 9.2e18), causing JSON decode failure
and "Challenge Failed" error. Changed seedValue to UInt64.

Added "Test Verification" button in Settings > Data section to
manually trigger auth challenge without needing a push notification.

+++ src/core/front_end/The Pendulum 2.0/The Pendulum 2.0/Core/Cipher/CipherAuthService.swift
@@ -42,7 +42,7 @@ final class CipherAuthService {
     }
 
     struct TransactionSeed: Codable {
-        let seedValue: Int
+        let seedValue: UInt64
         let seedHex: String
         let s1Position: Double
         let tier: Int
@@ -54,10 +54,7 @@ final class CipherAuthService {
         let difficulty: Double
         let timeLimit: Double?
         let parameters: [String: Double]
-        // parameters keys for pendulum:
-        //   balance_threshold, hold_duration, initial_angle,
-        //   mass, length, damping, gravity, spring_constant,
-        //   jiggle_intensity
+        // generated_maze is null for pendulum — ignored by Codable
     }
 
     struct ChallengeResponse: Codable {
